move MediaAtom model to common
also moves:
- `ThriftUtil`
- `JsonConversions`
- `MediaAtomImplicits`
