+++ app/model/commands/ActiveAssetCommand.scala
@@ -1,13 +1,13 @@
 package model.commands
 
 import com.gu.media.logging.Logging
+import com.gu.media.model.MediaAtom
 import com.gu.media.model.Platform.Youtube
+import com.gu.media.util.MediaAtomImplicits
 import com.gu.pandomainauth.model.{User => PandaUser}
 import data.DataStores
-import model.MediaAtom
 import model.commands.CommandExceptions._
 import util._
-import util.atom.MediaAtomImplicits
 
 case class ActiveAssetCommand(atomId: String, params: ActivateAssetRequest, stores: DataStores,
                               youTube: YouTube, user: PandaUser)
