+++ app/model/commands/AddPlutoProjectCommand.scala
@@ -1,9 +1,9 @@
 package model.commands
 
 import com.gu.media.logging.Logging
+import com.gu.media.model.MediaAtom
 import com.gu.pandomainauth.model.{User => PandaUser}
 import data.DataStores
-import model.MediaAtom
 import util.AWSConfig
 
 class AddPlutoProjectCommand(atomId: String, plutoId: String, override val stores: DataStores, user: PandaUser,
