--- app/views/MediaAtom/defaultEmbed.scala.html
@@ -1,24 +1,0 @@
-@import com.gu.media.model.{VideoAsset, YouTubeAsset, SelfHostedAsset}
-@(asset: Option[VideoAsset], posterUrl: Option[String])
-
-@defining((asset, posterUrl)) {
-  case (None, Some(poster)) => {
-    <img src="@poster"/>
-  }
-
-  case (Some(YouTubeAsset(id)), _) => {
-    <iframe frameborder="0" allowfullscreen="true" src="https://www.youtube.com/embed/@id?showinfo=0&rel=0"></iframe>
-  }
-
-  case (Some(SelfHostedAsset(sources)), maybePoster) => {
-    <video controls="controls" preload="metadata" @if(maybePoster.isDefined) {poster="@maybePoster.get"}>
-    @for(source <- sources) {
-      <source type="@source.mimeType" src="@source.src" />
-    }
-    </video>
-  }
-
-  case (None, None) => {
-    <div />
-  }
-}
