+++ common/src/main/scala/com/gu/media/model/MediaAtom.scala
@@ -1,10 +1,9 @@
-package model
+package com.gu.media.model
 
 import com.gu.contentatom.thrift.atom.media.{MediaAtom => ThriftMediaAtom, Metadata => ThriftMetadata}
 import com.gu.contentatom.thrift.{AtomData, Atom => ThriftAtom, AtomType => ThriftAtomType, Flags => ThriftFlags}
-import com.gu.media.model._
+import com.gu.media.util.MediaAtomImplicits
 import org.cvogt.play.json.Jsonx
-import util.atom.MediaAtomImplicits
 
 abstract class MediaAtomBase {
   //generic metadata
